First simple calculation of joint values

+++ app/jointPoseGetter/JointPoseGetter.h
@@ -2,7 +2,7 @@
 #include <vector>
 #include <eigen3/Eigen/Dense>
 
-typedef Eigen::Matrix<float, 6, 6> Matrix6f;
+typedef Eigen::Matrix<double, 6, 6> Matrix6d;
 
 
 class JointPoseGetter
@@ -12,8 +12,8 @@ private:
 public:
 	JointPoseGetter(/* args */);
 	~JointPoseGetter();
-    Matrix6f jacobian(double q1, double q2, double q3, double q4, double q5, double q6);
-    Matrix6f jacobianInverse(double q1, double q2, double q3, double q4, double q5, double q6);
+    Matrix6d jacobian(double q1, double q2, double q3, double q4, double q5, double q6);
+    Matrix6d jacobianInverse(double q1, double q2, double q3, double q4, double q5, double q6);
 
 private:
 };
